Implements SingleAccount and MultipleAccountPublicClientApplication
