+++ msal/src/main/java/com/microsoft/identity/client/PublicClientApplication.java
@@ -27,8 +27,6 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.os.Handler;
-import android.os.Looper;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.annotation.VisibleForTesting;
@@ -54,11 +52,9 @@ import com.microsoft.identity.common.internal.authorities.Authority;
 import com.microsoft.identity.common.internal.authorities.AuthorityDeserializer;
 import com.microsoft.identity.common.internal.authorities.AzureActiveDirectoryAudience;
 import com.microsoft.identity.common.internal.authorities.AzureActiveDirectoryAudienceDeserializer;
-import com.microsoft.identity.common.internal.authorities.AzureActiveDirectoryAuthority;
 import com.microsoft.identity.common.internal.cache.CacheKeyValueDelegate;
 import com.microsoft.identity.common.internal.cache.IAccountCredentialCache;
 import com.microsoft.identity.common.internal.cache.ICacheKeyValueDelegate;
-import com.microsoft.identity.common.internal.cache.IShareSingleSignOnState;
 import com.microsoft.identity.common.internal.cache.ISharedPreferencesFileManager;
 import com.microsoft.identity.common.internal.cache.MicrosoftStsAccountCredentialAdapter;
 import com.microsoft.identity.common.internal.cache.MsalOAuth2TokenCache;
@@ -68,9 +64,6 @@ import com.microsoft.identity.common.internal.controllers.ApiDispatcher;
 import com.microsoft.identity.common.internal.controllers.InteractiveTokenCommand;
 import com.microsoft.identity.common.internal.controllers.TokenCommand;
 import com.microsoft.identity.common.internal.dto.AccountRecord;
-import com.microsoft.identity.common.internal.migration.AdalMigrationAdapter;
-import com.microsoft.identity.common.internal.migration.TokenMigrationCallback;
-import com.microsoft.identity.common.internal.migration.TokenMigrationUtility;
 import com.microsoft.identity.common.internal.providers.microsoft.MicrosoftAccount;
 import com.microsoft.identity.common.internal.providers.microsoft.MicrosoftRefreshToken;
 import com.microsoft.identity.common.internal.providers.microsoft.azureactivedirectory.AzureActiveDirectory;
@@ -91,11 +84,8 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import static com.microsoft.identity.common.internal.cache.SharedPreferencesAccountCredentialCache.DEFAULT_ACCOUNT_CREDENTIAL_SHARED_PREFERENCES;
 
@@ -145,13 +135,13 @@ import static com.microsoft.identity.common.internal.cache.SharedPreferencesAcco
  * </p>
  * </p>
  */
-public class PublicClientApplication implements IPublicClientApplication, IMultipleAccountPublicClientApplication {
+public class PublicClientApplication implements IPublicClientApplication {
     private static final String TAG = PublicClientApplication.class.getSimpleName();
 
     private static final String INTERNET_PERMISSION = "android.permission.INTERNET";
     private static final String ACCESS_NETWORK_STATE_PERMISSION = "android.permission.ACCESS_NETWORK_STATE";
 
-    private PublicClientApplicationConfiguration mPublicClientConfiguration;
+    protected PublicClientApplicationConfiguration mPublicClientConfiguration;
 
     /**
      * {@link PublicClientApplication#create(Context, int, ApplicationCreatedListener)} will read the client id and other configuration settings from the
@@ -167,6 +157,8 @@ public class PublicClientApplication implements IPublicClientApplication, IMulti
      *                             </p>
      * @param configFileResourceId The resource ID of the raw file containing the JSON configuration for the PublicClientApplication
      * @see <a href="https://developer.android.com/guide/topics/resources/providing-resources">Android app resource overview</a>
+     *
+     * @param listener a callback to be invoked when the object is successfully created.
      * <p>
      * For more information on the schema of the MSAL config json please
      * @see <a href="https://github.com/AzureAD/microsoft-authentication-library-for-android/wiki">MSAL Github Wiki</a>
@@ -232,15 +224,14 @@ public class PublicClientApplication implements IPublicClientApplication, IMulti
         if (MsalUtils.isEmpty(clientId)) {
             throw new IllegalArgumentException("client id is empty or null");
         }
+
         new BrokerMsalController().getBrokerAccountMode(context, new BrokerAccountModeCallback() {
             @Override
             public void onGetMode(String mode) {
                 if (AuthenticationConstants.Broker.BROKER_ACCOUNT_MODE_SINGLE_ACCOUNT.equalsIgnoreCase(mode)) {
-                    // TODO: return SingleAccountPublicClientApplication
-                    listener.onCreated(new PublicClientApplication(context, clientId));
+                    listener.onCreated(new SingleAccountPublicClientApplication(context, clientId));
                 } else {
-                    // TODO: return MultipleAccountPublicClientApplication
-                    listener.onCreated(new PublicClientApplication(context, clientId));
+                    listener.onCreated(new MultipleAccountPublicClientApplication(context, clientId));
                 }
             }
         });
@@ -281,11 +272,9 @@ public class PublicClientApplication implements IPublicClientApplication, IMulti
             @Override
             public void onGetMode(String mode) {
                 if (AuthenticationConstants.Broker.BROKER_ACCOUNT_MODE_SINGLE_ACCOUNT.equalsIgnoreCase(mode)) {
-                    // TODO: return SingleAccountPublicClientApplication
-                    listener.onCreated(new PublicClientApplication(context, clientId, authority));
+                    listener.onCreated(new SingleAccountPublicClientApplication(context, clientId, authority));
                 } else {
-                    // TODO: return MultipleAccountPublicClientApplication
-                    listener.onCreated(new PublicClientApplication(context, clientId, authority));
+                    listener.onCreated(new MultipleAccountPublicClientApplication(context, clientId, authority));
                 }
             }
         });
@@ -298,11 +287,9 @@ public class PublicClientApplication implements IPublicClientApplication, IMulti
             @Override
             public void onGetMode(String mode) {
                 if (AuthenticationConstants.Broker.BROKER_ACCOUNT_MODE_SINGLE_ACCOUNT.equalsIgnoreCase(mode)) {
-                    // TODO: return SingleAccountPublicClientApplication
-                    listener.onCreated(new PublicClientApplication(context, developerConfig));
+                    listener.onCreated(new SingleAccountPublicClientApplication(context, developerConfig));
                 } else {
-                    // TODO: return MultipleAccountPublicClientApplication
-                    listener.onCreated(new PublicClientApplication(context,  developerConfig));
+                    listener.onCreated(new MultipleAccountPublicClientApplication(context,  developerConfig));
                 }
             }
         });
@@ -396,259 +383,8 @@ public class PublicClientApplication implements IPublicClientApplication, IMulti
         return mPublicClientConfiguration;
     }
 
-    @Override
-    public void getAccounts(@NonNull final AccountsLoadedCallback callback) {
-        ApiDispatcher.initializeDiagnosticContext();
-        final String methodName = ":getAccounts";
-        final List<AccountRecord> accounts = getLocalAccounts();
-
-        final Handler handler;
-
-        if (null != Looper.myLooper() && Looper.getMainLooper() != Looper.myLooper()) {
-            handler = new Handler(Looper.myLooper());
-        } else {
-            handler = new Handler(Looper.getMainLooper());
-        }
-
-        if (accounts.isEmpty()) {
-            // Create the SharedPreferencesFileManager for the legacy accounts/credentials
-            final IStorageHelper storageHelper = new StorageHelper(mPublicClientConfiguration.getAppContext());
-            final ISharedPreferencesFileManager sharedPreferencesFileManager =
-                    new SharedPreferencesFileManager(
-                            mPublicClientConfiguration.getAppContext(),
-                            "com.microsoft.aad.adal.cache",
-                            storageHelper
-                    );
-
-            // Load the old TokenCacheItems as key/value JSON
-            final Map<String, String> credentials = sharedPreferencesFileManager.getAll();
-
-            final Map<String, String> redirects = new HashMap<>();
-            redirects.put(
-                    mPublicClientConfiguration.mClientId, // Our client id
-                    mPublicClientConfiguration.mRedirectUri // Our redirect uri
-            );
-
-            new TokenMigrationUtility<MicrosoftAccount, MicrosoftRefreshToken>()._import(
-                    new AdalMigrationAdapter(
-                            mPublicClientConfiguration.getAppContext(),
-                            redirects,
-                            false
-                    ),
-                    credentials,
-                    (IShareSingleSignOnState<MicrosoftAccount, MicrosoftRefreshToken>) mPublicClientConfiguration.getOAuth2TokenCache(),
-                    new TokenMigrationCallback() {
-                        @Override
-                        public void onMigrationFinished(int numberOfAccountsMigrated) {
-                            final String extendedMethodName = ":onMigrationFinished";
-                            com.microsoft.identity.common.internal.logging.Logger.info(
-                                    TAG + methodName + extendedMethodName,
-                                    "Migrated [" + numberOfAccountsMigrated + "] accounts"
-                            );
-                            // Merge migrated accounts with broker or local accounts.
-                            if (MSALControllerFactory.brokerEligible(
-                                    mPublicClientConfiguration.getAppContext(),
-                                    mPublicClientConfiguration.getDefaultAuthority(),
-                                    mPublicClientConfiguration)) {
-                                postBrokerAndLocalAccountsResult(handler, callback);
-                            } else {
-                                postLocalAccountsResult(handler, callback);
-                            }
-                        }
-                    }
-            );
-        } else {
-            // The cache contains items - mark migration as complete
-            new AdalMigrationAdapter(
-                    mPublicClientConfiguration.getAppContext(),
-                    null, // unused for this path
-                    false
-            ).setMigrationStatus(true);
-
-            if (MSALControllerFactory.brokerEligible(
-                    mPublicClientConfiguration.getAppContext(),
-                    mPublicClientConfiguration.getDefaultAuthority(),
-                    mPublicClientConfiguration)) {
-                postBrokerAndLocalAccountsResult(handler, callback);
-            } else {
-                postLocalAccountsResult(handler, callback);
-            }
-        }
-    }
-
-    /**
-     * Helper method which returns all the local accounts using {@link AccountsLoadedCallback}
-     * @param handler : handler to post
-     * @param callback: AccountsLoadedCallback
-     */
-    private void postLocalAccountsResult(final Handler handler, final AccountsLoadedCallback callback) {
-
-        handler.post(new Runnable() {
-            @Override
-            public void run() {
-                List<IAccount> accountsToReturn = new ArrayList<>();
-                for (AccountRecord accountRecord : getLocalAccounts()) {
-                    accountsToReturn.add(AccountAdapter.adapt(accountRecord));
-                }
-
-                callback.onAccountsLoaded(accountsToReturn);
-            }
-        });
-    }
-
-    /**
-     * Helper method which returns both broker and local accounts using {@link AccountsLoadedCallback}
-     * @param handler : handler to post
-     * @param callback: AccountsLoadedCallback
-     */
-    private void postBrokerAndLocalAccountsResult(final Handler handler, final AccountsLoadedCallback callback) {
-
-        final String methodName = ":postBrokerAndLocalAccountsResult";
-
-        new BrokerMsalController().getBrokerAccounts(
-                mPublicClientConfiguration,
-                new BrokerAccountsLoadedCallback() {
-                    @Override
-                    public void onAccountsLoaded(final List<AccountRecord> accountRecords) {
-                        com.microsoft.identity.common.internal.logging.Logger.verbose(
-                                TAG + methodName,
-                                "Accounts loaded from broker "
-                                        + (accountRecords == null ? 0 : accountRecords.size())
-                        );
-
-                        // merge account
-                        final List<IAccount> accountList = new ArrayList<>();
-                        final List<AccountRecord> accountRecordList = new ArrayList<>();
-
-                        if (accountRecords != null) {
-                            //Add broker accounts
-                            accountRecordList.addAll(accountRecords);
-                        }
-
-                        //Add local accounts
-                        accountRecordList.addAll(getLocalAccounts());
-
-                        if (accountRecordList.size() > 0) {
-                            for (AccountRecord accountRecord : accountRecordList) {
-                                accountList.add(AccountAdapter.adapt(accountRecord));
-                            }
-                        }
-                        handler.post(new Runnable() {
-                            @Override
-                            public void run() {
-                                callback.onAccountsLoaded(accountList);
-                            }
-                        });
-                    }
-                });
-    }
-
-    /**
-     * Returns a List of {@link IAccount} objects for which this application has RefreshTokens.
-     *
-     * @return An immutable List of IAccount objects - empty if no IAccounts exist.
-     */
-    private List<AccountRecord> getLocalAccounts() {
-        // Grab the Accounts from the common cache
-        final List<AccountRecord> accountsInCache =
-                mPublicClientConfiguration
-                        .getOAuth2TokenCache()
-                        .getAccounts(
-                                null, // * wildcard
-                                mPublicClientConfiguration.getClientId()
-                        );
-
-        return accountsInCache;
-    }
-
-    @Override
     @Nullable
-    public IAccount getAccount(@NonNull final String homeAccountIdentifier,
-                               @Nullable final String authority) {
-        final String methodName = ":getAccount";
-
-        ApiDispatcher.initializeDiagnosticContext();
-
-        String realm = StringUtil.getTenantInfo(homeAccountIdentifier).second;
-
-        Authority authorityObj = Authority.getAuthorityFromAuthorityUrl(authority);
-
-        if (authorityObj instanceof AzureActiveDirectoryAuthority) {
-            final AzureActiveDirectoryAuthority aadAuthority = (AzureActiveDirectoryAuthority) authorityObj;
-            final AzureActiveDirectoryAudience audience = aadAuthority.getAudience();
-            realm = audience.getTenantId();
-        } else {
-            com.microsoft.identity.common.internal.logging.Logger.warn(
-                    TAG + methodName,
-                    "Provided authority was not AAD - defaulting to parsed home_account_id"
-            );
-        }
-
-        AccountRecord accountToReturn = null;
-
-        if (null != realm) {
-            accountToReturn = AccountAdapter.getAccountInternal(
-                    mPublicClientConfiguration.getClientId(),
-                    mPublicClientConfiguration.getOAuth2TokenCache(),
-                    homeAccountIdentifier,
-                    realm
-            );
-        } else {
-            com.microsoft.identity.common.internal.logging.Logger.warn(
-                    TAG + methodName,
-                    "Realm could not be resolved. Returning null."
-            );
-        }
-
-        return null == accountToReturn ? null : AccountAdapter.adapt(accountToReturn);
-    }
-
-    @Override
-    public void removeAccount(@Nullable final IAccount account, final AccountRemovedListener callback) {
-        ApiDispatcher.initializeDiagnosticContext();
-        if (null == account
-                || null == account.getHomeAccountIdentifier()
-                || StringUtil.isEmpty(account.getHomeAccountIdentifier().getIdentifier())) {
-            com.microsoft.identity.common.internal.logging.Logger.warn(
-                    TAG,
-                    "Requisite IAccount or IAccount fields were null. Insufficient criteria to remove IAccount."
-            );
-
-            callback.onAccountRemoved(false);
-        }
-
-        // FEATURE SWITCH: Set to false to allow deleting Accounts in a tenant-specific way.
-        final boolean deleteAccountsInAllTenants = true;
-
-        final String realm = deleteAccountsInAllTenants ? null : getRealm(account);
-
-        final boolean localRemoveAccountSuccess = !mPublicClientConfiguration
-                .getOAuth2TokenCache()
-                .removeAccount(
-                        account.getEnvironment(),
-                        mPublicClientConfiguration.getClientId(),
-                        account.getHomeAccountIdentifier().getIdentifier(),
-                        realm
-                ).isEmpty();
-
-        if (MSALControllerFactory.brokerEligible(
-                mPublicClientConfiguration.getAppContext(),
-                mPublicClientConfiguration.getDefaultAuthority(),
-                mPublicClientConfiguration)) {
-
-            //Remove the account from Broker
-            new BrokerMsalController().removeBrokerAccount(
-                    account,
-                    mPublicClientConfiguration,
-                    callback
-            );
-        } else {
-            callback.onAccountRemoved(localRemoveAccountSuccess);
-        }
-    }
-
-    @Nullable
-    private static String getRealm(@NonNull IAccount account) {
+    protected static String getRealm(@NonNull IAccount account) {
         String realm = null;
 
         if (null != account.getAccountIdentifier() // This is an AAD account w/ tenant info
